+++ src/ui-spa/src/common/utils/getSuspectJourneyRoutes.ts
@@ -1,0 +1,108 @@
+import { type SuspectAdditionalDetailValue } from "../reducers/caseRegistrationReducer";
+
+const routeSequence: {
+  route: string;
+  description: SuspectAdditionalDetailValue | "Add suspect";
+}[] = [
+  {
+    route: "add-suspect",
+    description: "Add suspect",
+  },
+  {
+    route: "suspect-dob",
+    description: "Date of Birth",
+  },
+  {
+    route: "suspect-gender",
+    description: "Gender",
+  },
+  {
+    route: "suspect-disability",
+    description: "Disability",
+  },
+  {
+    route: "suspect-religion",
+    description: "Religion",
+  },
+  {
+    route: "suspect-ethnicity",
+    description: "Ethnicity",
+  },
+  {
+    route: "suspect-add-aliases",
+    description: "Alias details",
+  },
+  {
+    route: "suspect-sdo",
+    description: "Serious dangerous offender (SDO)",
+  },
+  {
+    route: "suspect-asn",
+    description: "Arrest summons number (ASN)",
+  },
+  {
+    route: "suspect-offender",
+    description: "Type of offender",
+  },
+];
+
+export const getNextSuspectJourneyRoute = (
+  currentRoute: string,
+  suspectAdditionalDetailsCheckboxes: SuspectAdditionalDetailValue[],
+  currentSuspectIndex: number,
+): string => {
+  if (!suspectAdditionalDetailsCheckboxes.length) return "/suspect-summary";
+
+  const shouldInclude = (description: SuspectAdditionalDetailValue) => {
+    return suspectAdditionalDetailsCheckboxes.includes(description);
+  };
+
+  const filteredSequence = routeSequence.filter((item) => {
+    if (item.description !== "Add suspect") {
+      return shouldInclude(item.description);
+    }
+    return false;
+  });
+
+  console.log("filteredSequence", filteredSequence);
+  const currentIndex = filteredSequence.findIndex(
+    (item) => item.route === currentRoute,
+  );
+
+  if (filteredSequence[currentIndex + 1]?.route) {
+    return `/case-registration/suspect-${currentSuspectIndex}/${filteredSequence[currentIndex + 1].route}`;
+  }
+
+  return `/case-registration/suspect-summary`;
+};
+
+export const getPreviousSuspectJourneyRoute = (
+  currentRoute: string,
+  suspectAdditionalDetailsCheckboxes: SuspectAdditionalDetailValue[],
+  currentSuspectIndex: number,
+): string => {
+  const shouldInclude = (description: SuspectAdditionalDetailValue) => {
+    return suspectAdditionalDetailsCheckboxes.includes(description);
+  };
+
+  const filteredSequence = routeSequence.filter((item) => {
+    if (item.description !== "Add suspect") {
+      return shouldInclude(item.description);
+    }
+    return false;
+  });
+
+  const currentIndex = filteredSequence.findIndex(
+    (item) => item.route === currentRoute,
+  );
+
+  // if (currentIndex === 0) {
+  //   return `/case-registration/suspect-${currentSuspectIndex}/add-suspect`;
+  // }
+
+  if (filteredSequence[currentIndex - 1]?.route) {
+    return `/case-registration/suspect-${currentSuspectIndex}/${filteredSequence[currentIndex - 1].route}`;
+  }
+
+  return `/case-registration/suspect-${currentSuspectIndex}/add-suspect`;
+};
